Bin atac fragments

Convenience wrapper to bin fragments of a given size and save them as .mtx files

Usage

bin_atac_frags(
  ArrowFiles,
  bins,
  blacklist = NULL,
  outdir,
  bin_name = prettyMb(getmode(width(bins))),
  ncores = 1,
  bpparams = BiocParallel::MulticoreParam(workers = ncores, progressbar = TRUE),
  overwrite = FALSE,
  return_mat = FALSE,
  ...
)

Arguments

ArrowFiles

List or vector of ArrowFile paths

bins

Bins GRanges object

outdir

Optional: Directory to write the .mtx, barcodes, and bins files

bin_name

Name of the bins (e.g. '10Mb', '500Kb', 'chr_arm'). If not provided is automatically detected based on binwidth.

ncores

Number of cores to use

bpparams

Options to BPPARAM to enable multithreading

overwrite

Logical. Overwrite previously existing results (default = FALSE)

return_mat

Logical. Return the binned depth matrix (default = FALSE)

...

Additional parameters passed to bplapply

Value

If return_mat=TRUE, returns a sparse binned depth matrix. Otherwise returns NULL