Cluster using Seurat

Cluster using Seurat

Usage

cluster_seurat(
  sce,
  assay_name,
  do.scale = FALSE,
  do.center = FALSE,
  algorithm = 1,
  resolution = 0.8,
  n.neighbors = 10,
  npcs.pca = 50,
  features.pca = "all",
  nvar.features = NULL,
  dims = 1:npcs.pca,
  k.param = 20,
  suffix = "",
  PCA_name = paste0("PCA", suffix),
  UMAP_name = paste0("UMAP", suffix),
  cluster_name = paste0("clusters", suffix),
  umap.metric = "correlation",
  annoy.metric = "cosine",
  verbose = TRUE
)

Arguments

sce

SingleCellExperiment object

assay_name

Assay name. Can provide two assay names to perform joint clustering across both

do.scale

scale

do.center

center

algorithm

clustering algorithm

resolution

clustering resolution

n.neighbors

neighbors for umap

npcs.pca

Total Number of PCs to compute and store (50 by default)

features.pca

One of 'all', 'variable', or a vector of features to include in dimensionality reduction. Defaults to 'all'.

nvar.features

Number of variable features if features.pca='variable'

dims

Number of reduced dimensions to use for FindNeighbors and UMAP

k.param

Defines k for the k-nearest neighbor algorithm

suffix

Suffix name to add to the PCA, UMAP, and clusters

PCA_name

Name to store PCA dimred

UMAP_name

Name to store UMAP dimred

cluster_name

Name to store seurat clusters

umap.metric

Metric for Seurat::RunUMAP

annoy.metric

Metric for Seurat::FindNeighbors

verbose

Message verbosity

Value

SingleCellExperiment obj