Perform UMAP and clustering

Clusters a copy number matrix using hdbscan.

Usage

perform_umap_clustering(
  cn_matrix,
  n_neighbors = 10,
  min_dist = 0.1,
  minPts = 30,
  scale = c("none", "cells", "bins", "both"),
  log2 = FALSE,
  seed = 3,
  metric = "correlation",
  verbose = TRUE
)

Arguments

cn_matrix

Copy number matrix with cells in columns and bins in rows

n_neighbors

The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100.

min_dist

The effective minimum distance between embedded points. Smaller values will result in a more clustered/clumped embedding where nearby points on the manifold are drawn closer together, while larger values will result on a more even dispersal of points. The value should be set relative to the spread value, which determines the scale at which embedded points will be spread out.

scale

One of 'cells', 'bins', 'both' or 'none'. Determines what kind of scaling is done.

log2

Logical: Log2 transform the matrix prior to plotting

metric

Type of distance metric to use to find nearest neighbors. One of:

"euclidean" (the default)

"cosine"

"manhattan"

"hamming"

"correlation" (a distance based on the Pearson correlation)

"categorical" (see below)

Only applies if nn_method = "annoy" (for nn_method = "fnn", the distance metric is always "euclidean").

If X is a data frame or matrix, then multiple metrics can be specified, by passing a list to this argument, where the name of each item in the list is one of the metric names above. The value of each list item should be a vector giving the names or integer ids of the columns to be included in a calculation, e.g. metric = list(euclidean = 1:4, manhattan = 5:10).

Each metric calculation results in a separate fuzzy simplicial set, which are intersected together to produce the final set. Metric names can be repeated. Because non-numeric columns are removed from the data frame, it is safer to use column names than integer ids.

Factor columns can also be used by specifying the metric name "categorical". Factor columns are treated different from numeric columns and although multiple factor columns can be specified in a vector, each factor column specified is processed individually. If you specify a non-factor column, it will be coerced to a factor.

For a given data block, you may override the pca and pca_center arguments for that block, by providing a list with one unnamed item containing the column names or ids, and then any of the pca or pca_center overrides as named items, e.g. metric = list(euclidean = 1:4, manhattan = list(5:10, pca_center = FALSE)). This exists to allow mixed binary and real-valued data to be included and to have PCA applied to both, but with centering applied only to the real-valued data (it is typical not to apply centering to binary data before PCA is applied).

verbose

Logical: Message verbosity

Value

A list with the clustering results

clustering

Data frame of cell and cluster identities

hdbscanresults

Results from dbscan::hdbscan

umapresults

Results from uwot::umap

tree

Results from ape::as.phylo on hdbscanresults$hc

Examples

clust_results <- perform_umap_clustering(cn_matrix = assay(test_sce, "counts"))
#> 13:40:47 UMAP embedding parameters a = 1.577 b = 0.8951
#> 13:40:47 Read 75 rows and found 310 numeric columns
#> 13:40:47 Using Annoy for neighbor search, n_neighbors = 10
#> 13:40:47 Annoy build: subtracting row means for correlation
#> 13:40:47 Building Annoy index with metric = correlation, n_trees = 50
#> 0%   10   20   30   40   50   60   70   80   90   100%
#> [----|----|----|----|----|----|----|----|----|----|
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> |
#> 13:40:47 Annoy search: subtracting row means for correlation
#> 13:40:47 Writing NN index file to temp file /tmp/RtmpePnXJq/file63f45968b36f
#> 13:40:47 Searching Annoy index using 36 threads, search_k = 1000
#> 13:40:47 Annoy recall = 100%
#> 13:40:48 Commencing smooth kNN distance calibration using 36 threads
#> 13:40:49 Initializing from normalized Laplacian + noise
#> 13:40:49 Commencing optimization for 500 epochs, with 1198 positive edges
#> 13:40:50 Optimization finished
#>   Cluster 0: 75