scATAC CBS CNV segmentation

scATAC CBS CNV segmentation

Usage

segment_cnv(
  sce,
  assay_name,
  new_assay = paste(assay_name, "segment", sep = "_"),
  alpha = 0.2,
  nperm = 10,
  min.width = 2,
  undo.splits = "none",
  verbose = 0,
  bpparam = BiocParallel::SerialParam(),
  ...
)

Arguments

sce

A SingleCellExperiment object

assay_name

Name of the assay to segment

new_assay

Name of the new assay

alpha

significance levels for the test to accept change-points.

nperm

number of permutations used for p-value computation.

min.width

the minimum number of markers for a changed segment. The default is 2 but can be made larger. Maximum possible value is set at 5 since arbitrary widths can have the undesirable effect of incorrect change-points when a true signal of narrow widths exists.

undo.splits

A character string specifying how change-points are to be undone, if at all. Default is "none". Other choices are "prune", which uses a sum of squares criterion, and "sdundo", which undoes splits that are not at least this many SDs apart.

verbose

Verbosity

bpparam

BiocParallel Parameters

...

Additional parameters to pass to DNAcopy::segment

Value

A SingleCellExperiment object