Bin atac fragments

Convenience wrapper to bin fragments of a given size and save them as .mtx files

Usage

bin_atac_frags(
  sample_id,
  fragment_file,
  cells = NULL,
  bins,
  bin_name = NULL,
  blacklist = NULL,
  outdir,
  ncores = 1,
  bpparam = BiocParallel::SerialParam(),
  overwrite = FALSE,
  return_mat = FALSE
)

Arguments

sample_id

Sample id

fragment_file

Fragments file

cells

A vector of cell barcodes. If not provided will use all barcodes

bins

Genomic ranges object with bins to use

bin_name

bin_name Name of the bins (e.g. '10Mb', '500Kb', 'chr_arm'). If not provided is automatically detected based on binwidth.

blacklist

Genome blacklist regions to filter against

outdir

Output directory

ncores

Number of cores

bpparam

BiocParallel Parameters

overwrite

Whether to overwrite previous binned counts

return_mat

Logical. Return the binned depth matrix (default = FALSE)

Value

If return_mat=TRUE, returns a sparse binned depth matrix. Otherwise returns NULL, but will save the results to a 10X style market matrix outdir