Flag ideal bins

is_ideal_bin will apply a set of bin-wise filters, based on high count outliers, high or low gc outliers, minimum read counts, minimum mappability, or maximum allowable frequency of N bases per bin.

Usage

is_ideal_bin(
  counts,
  gc,
  n_freq,
  map = NULL,
  min_reads = 0,
  max_N_freq = 0.05,
  reads_outlier = 0.01,
  gc_outlier = 0.001,
  min_map = 0.9
)

Arguments

counts

Vector of bin counts for single cell

gc

Vector of gc content

n_freq

Vector of bin N frequency (proportion of N bases in a bin)

map

Vector of bin mappability

min_reads

Minimum number of reads to consider a bin

max_N_freq

Maximum allowable frequency of N bases to consider a bin. Range (0, 1)

reads_outlier

Flag bins with reads in the top quantile given by this value. Range (0, 1)

gc_outlier

Flag bins with GC content in the top and bottom quantule given by this value. Range (0, 1)

min_map

Minimum allowable mappability score for a bin. Range (0, 1)

Value

A dataframe of two columns meet the valid or ideal criteria